[FIX] Columns with JDBC_TYPE_BIGINT is being returned as having DATA_TYPE -2 (BINARY) instead of -5 in SqlColumns (#44)
* ODBC was returing  the wrong data_type
type conversion from int8 to uint8 was causing the isse

* adding schema and table name in SQLCOLUMNS test

* changing int8 to int16

+++ src/odbc/include/ignite/odbc/meta/column_meta.h
@@ -205,7 +205,7 @@ class ColumnMeta {
    * Get data type.
    * @return Data type.
    */
-  int8_t GetDataType() const {
+  int16_t GetDataType() const {
     return dataType;
   }
